[WIP] Feature/open xpack (#172)
* existing unit tests back to green

* Show new installation path on notice step, also show notice step during
patch upgrades.

* fix failing unit tests after patch upgrades show notice screen

* NoticeView now includes information about x-pack being included when upgrading from <= 6.3.0

* Made sure read more button only shows for minor and major version changes

* x-pack step touchups

* bootstrap password task was checking x-pack in plugin list still

* fixed failing unit test that asserted that by default we do not write a self generated license type in the elasticsearch yaml

* Read more about opening xpack links now leads user to landing page on elastic.co, make sure we seed a random Bootstrap Password that is not empty

* update wxs file

* always run SetBootstrapPasswordTask when XPackModel is relevant

* Add ability to launch vagrant with gui and not destroy on test finish

This commit adds the ability to control vagrant when running integration tests by
supplying command line switches to the integrate build task:

-gui : runs vagrant with a gui
-nodestroy : does not destroy the vagrant box when the test(s) finish

Both options are only supported for local vagrant boxes.

* Create temp directory if not exists

The temporary directory must exist when the elasticsearch process starts up.
If it does not, the process fails indicating it does not exist.

* make sure we preserve comments/blank lines and jdk dependant settings when editing the jvm.options file

* Provision machine with Java 9

Comment out provisioning for the moment until choosing between java8 and java9 is implemented

* Update integration tests to handle 6.3.0

This commit updates the integration tests to handle installation of 6.3.0+, which includes X-PAck basic by default.

* Don't try to install X-Pack for any 6.3.0 preleases

This commit updates the conditional check for whether to install X-Pack
plugins to also consider 6.3.0 prereleases

* Keep all files in bin directory except shell scripts and specific bat files

This commit updates the logic for files that are pruned from the bin directory.
There are now many more batch files that must be included, so take an inclusive
approach excepting files known to not be needed.

* Explicitly set XPackSecurtiyEnabled true for Trial license

This commit sets XpackSecurityEnabled to true when a trial license is selected. An unset
xpack.security.enabled value in config used to imply enabled in previous versions, however
going forward, will need to be explicitly set to true.

Always persist a value for XpackSecurityEnabled to config.

* Install plugins from local zip in installation tests

This commit adds the ability to specify a list of plugins to install from a local zip within integration tests.
A list of plugins to copy to build/out can be supplied with -plugins:<comma separated plugin names>

For example

    build.bat integrate es r:6.2.3,6.3.0-SNAPSHOT local * -plugins:ingest-geoip,ingest-attachment

At build time, the build/in directory will be searched for each specified version and plugin combination, and
if found, will be copied to the build/out directory to make available to integration tests.

When an integration test that will install plugins runs, the out directory is checked for a plugin zip of the
name and version and if it exists, the installer will use it.

* Remove link capitalization

* Show notice view when installing prerelease version

This commit updates the header when installing a prerelease version vs. upgrading to a prerelease version

* Fix erroneous integration test

This commit fixes the integration to make the semantic version comparisons between
the versions in question for the Describe block.

Conflicts:
	src/Installer/Elastic.Installer.Msi/_Generated/elasticsearch.wxs
	src/ProcessHosts/Elastic.ProcessHosts.Elasticsearch/Properties/AssemblyInfo.cs

+++ src/Tests/Elastic.Installer.Integration.Tests/Common/Utils.ps1
@@ -223,7 +223,7 @@ function Add-Cygpath() {
     }
 }
 
-function Invoke-IntegrationTestsOnLocal($Location, $Version, $PreviousVersions, $VagrantProvider) {
+function Invoke-IntegrationTestsOnLocal($Location, $Version, $PreviousVersions, $VagrantProvider, $VagrantDestroy) {
     cd $Location  
     $testDirName = Split-Path $Location -Leaf
     vagrant destroy local -f
@@ -233,7 +233,9 @@ function Invoke-IntegrationTestsOnLocal($Location, $Version, $PreviousVersions, 
 		vagrant powershell local -c "C:\common\PesterBootstrap.ps1 -Version '$Version' -PreviousVersions @($($($PreviousVersions | ForEach-Object { "'$_'" }) -join ",")) -TestDirName '$testDirName'"
 	}
 	finally {
-		vagrant destroy local -f
+		if ($VagrantDestroy) {
+			vagrant destroy local -f
+		}
 	}
 }
 
@@ -340,7 +342,7 @@ function Invoke-IntegrationTestsOnQuickAzure($Location, $Version, $PreviousVersi
 	}
 }
 
-function Invoke-IntegrationTests($CurrentDir, $TestDirs, $VagrantProvider, $Version, $PreviousVersions) {
+function Invoke-IntegrationTests($CurrentDir, $TestDirs, $VagrantProvider, $Version, $PreviousVersions, $Gui, $VagrantDestroy) {
 
 	# run all tests on one vagrant box
 	if ($VagrantProvider -eq "quick-azure") {
@@ -391,7 +393,11 @@ function Invoke-IntegrationTests($CurrentDir, $TestDirs, $VagrantProvider, $Vers
 			Copy-Item "$CurrentDir\common\Vagrantfile" -Destination $dir -Force
 
 			if ($VagrantProvider -eq "local") {
-				Invoke-IntegrationTestsOnLocal -Location $dir -Version $Version -PreviousVersions $PreviousVersions 
+				if ($Gui) {
+					ReplaceInFile -File "$dir\Vagrantfile" -Replacements @{ "#vb.gui = true" = "vb.gui = true" }
+				}
+
+				Invoke-IntegrationTestsOnLocal -Location $dir -Version $Version -PreviousVersions $PreviousVersions -VagrantDestroy:$VagrantDestroy
 			} 
 			else {
 				Invoke-IntegrationTestsOnAzure -Location $dir -Version $Version -PreviousVersions $PreviousVersions 
@@ -478,6 +484,29 @@ function Invoke-SilentInstall {
 		$Upgrade
     )
 
+	# if there are local zips of the plugins in build/out with the same version,
+	# use them for installation
+	if ($Exeargs) {
+		$Newargs = New-Object "System.Collections.ArrayList"
+		$Exeargs | %{
+			if ($_.StartsWith("PLUGINS=")) {
+				$plugins = $_.Replace("PLUGINS=", "").Split(',') | %{
+					$plugin = $_
+					$zip = ".\..\out\$plugin-$($Version.FullVersion).zip"
+					if (Test-Path $zip) {
+						(Get-Item $zip).FullName
+					} else {
+						$plugin
+					}
+				}
+				$Newargs.Add("PLUGINS=$($plugins -join ",")") | Out-Null
+			} else {
+				$Newargs.Add($_) | Out-Null
+			}		
+		}
+		$Exeargs = $Newargs
+	}
+
 	if ($Version.Source) {
         if (!$Exeargs) {
             $Exeargs = @("PLUGINSSTAGING=$($Version.Source)")
@@ -486,6 +515,7 @@ function Invoke-SilentInstall {
             $Exeargs.Add("PLUGINSSTAGING=$($Version.Source)") | Out-Null
         }
     }
+
 
     $QuotedArgs = Add-Quotes $Exeargs
     $Exe = Get-Installer -Version $Version
@@ -496,7 +526,7 @@ function Invoke-SilentInstall {
 		$logFile = "install.log"
 	}
 	$argumentList = "/i $Exe /qn /l*v $logFile $QuotedArgs"
-    log "running installer: msiexec.exe $argumentList"
+    log "running installer: msiexec.exe $argumentList" -Level Debug
     $ExitCode = (Start-Process C:\Windows\System32\msiexec.exe -ArgumentList $argumentList -Wait -PassThru).ExitCode
 
     if ($ExitCode) {
